Feature/graphql scalars

+++ src/main/java/com/youssef/GridPulse/domain/message/payload/SystemPayload.java
@@ -25,7 +25,7 @@ public record SystemPayload(
         int cpuUsage,       // percentage
         int memoryUsage,    // percentage
         int diskUsage,      // percentage
-        int temperature,    // in Celsius
+        Float temperature,    // in Celsius
         double latitude,    // GPS latitude
         double longitude    // GPS longitude
 ) {}
